requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
boto3==1.34.11
pytz==2023.3.post1
tabulate==0.9.0
python-mlb-statsapi